+++ .eslintrc.json
@@ -3,7 +3,7 @@
   "extends": ["eslint:recommended"],
   "overrides": [
     {
-      "files": ["**/*.?([cm])ts"],
+      "files": ["**/*.ts"],
       "extends": ["plugin:@typescript-eslint/recommended"],
       "plugins": ["eslint-plugin-tsdoc"],
       "rules": {
