--- src/index.mts
@@ -1,1 +1,0 @@
-export { fibonacciSequence } from "./sequence.mjs";
