+++ src/sequence.ts
